# docs/agent-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Silver Residence Agent Integration API
  version: 1.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Shared contract for the browser-facing Spring Boot Agent Gateway,
    LangGraph runs and RAG, and Spring Boot internal Agent Tools.
servers:
  - url: /
    description: Same-origin Spring Boot gateway and Agent Tool service
...
paths:
  /api/assistant/runs/stream:
    post:
      tags: [Assistant Gateway]
      summary: Create a browser-facing streaming Agent run
      operationId: streamAssistantRun
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/BrowserTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantRunRequest'
      responses:
        '200':
          description: Named SSE event stream
          headers:
            X-Trace-Id:
              schema: {type: string}
          content:
            text/event-stream:
              schema: {type: string}
              x-event-data-schema:
                $ref: '#/components/schemas/SseEnvelope'
              examples:
                hybrid:
                  externalValue: ./examples/agent-sse-hybrid.txt
        '400':
          $ref: '#/components/responses/SseFailure'
        '401':
          $ref: '#/components/responses/SseFailure'
        '503':
          $ref: '#/components/responses/SseFailure'
  /api/assistant/runs/{runId}/cancel:
    post:
      tags: [Assistant Gateway]
      summary: Cancel a browser-owned Agent run
      operationId: cancelAssistantRun
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RunIdPath'
        - $ref: '#/components/parameters/BrowserTraceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string, maxLength: 128}
              additionalProperties: false
      responses:
        '200':
          description: Current cancelled or terminal run status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStatusResponse'
        '401':
          $ref: '#/components/responses/JsonFailure'
        '503':
          $ref: '#/components/responses/JsonFailure'
  /api/assistant/runs/{runId}/events:
    get:
      tags: [Assistant Gateway]
      summary: Reattach to a browser-owned Agent run and replay later events
      operationId: replayAssistantRunEvents
      x-implementation-status: implemented
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RunIdPath'
        - $ref: '#/components/parameters/AfterSequence'
        - $ref: '#/components/parameters/BrowserTraceId'
      responses:
        '200':
          description: Events with sequence greater than afterSequence, followed by live events
          content:
            text/event-stream:
              schema: {type: string}
              x-event-data-schema:
                $ref: '#/components/schemas/SseEnvelope'
        '401':
          $ref: '#/components/responses/SseFailure'
        '404':
          $ref: '#/components/responses/SseFailure'
        '410':
          $ref: '#/components/responses/SseFailure'
...
  /api/v1/runs/stream:
    post:
      tags: [LangGraph Runs]
      summary: Create a streaming LangGraph run
      operationId: streamLangGraphRun
      servers:
        - url: '{langGraphBaseUrl}'
          description: Internal LangGraph service
          variables:
            langGraphBaseUrl:
              default: http://127.0.0.1:8000
      security:
        - serviceBearer: []
      parameters:
        - $ref: '#/components/parameters/InternalTraceId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LangGraphRunRequest'
      responses:
        '200':
          description: Versioned named SSE stream
          content:
            text/event-stream:
              schema: {type: string}
              x-event-data-schema:
                $ref: '#/components/schemas/SseEnvelope'
              examples:
                hybrid:
                  externalValue: ./examples/agent-sse-hybrid.txt
        '400':
          $ref: '#/components/responses/JsonFailure'
        '401':
          $ref: '#/components/responses/JsonFailure'
...
  /api/v1/runs/{runId}/events:
    get:
      tags: [LangGraph Runs]
      summary: Replay persisted events and continue the live stream
      operationId: replayLangGraphRunEvents
      x-implementation-status: required-for-production
      servers:
        - url: '{langGraphBaseUrl}'
          description: Internal LangGraph service
          variables:
            langGraphBaseUrl:
              default: http://127.0.0.1:8000
      security:
        - serviceBearer: []
      parameters:
        - $ref: '#/components/parameters/RunIdPath'
        - $ref: '#/components/parameters/AfterSequence'
        - $ref: '#/components/parameters/InternalTraceId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Events with sequence greater than afterSequence, followed by live events
          content:
            text/event-stream:
              schema: {type: string}
              x-event-data-schema:
                $ref: '#/components/schemas/SseEnvelope'
        '404':
          $ref: '#/components/responses/JsonFailure'
        '410':
          $ref: '#/components/responses/JsonFailure'
  /api/v1/runs/{runId}/cancel:
    post:
      tags: [LangGraph Runs]
      summary: Idempotently cancel a LangGraph run
      operationId: cancelLangGraphRun
      servers:
        - url: '{langGraphBaseUrl}'
          description: Internal LangGraph service
          variables:
            langGraphBaseUrl:
              default: http://127.0.0.1:8000
      security:
        - serviceBearer: []
      parameters:
        - $ref: '#/components/parameters/RunIdPath'
        - $ref: '#/components/parameters/InternalTraceId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: {type: string, maxLength: 128}
              required: [reason]
              additionalProperties: false
      responses:
        '200':
          description: Cancelled or existing terminal state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStatusResponse'
        '404':
          $ref: '#/components/responses/JsonFailure'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 access token
  parameters:
    BrowserTraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema: {type: string, maxLength: 128}
    InternalTraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
    AfterSequence:
      name: afterSequence
      in: query
      required: true
      description: Replay only events whose sequence is greater than this value.
      schema: {type: integer, minimum: 0}
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema: {type: string, maxLength: 128}
    UserId:
      name: X-User-Id
      in: header
      required: true
      schema: {type: string, maxLength: 128}
...
    RunIdPath:
      name: runId
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    JsonFailure:
      description: Contract failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ContractFailure'
    SseFailure:
      description: Failure before a Run is accepted; no run.started event is emitted
      content:
        text/event-stream:
          schema: {type: string}
          x-event-data-schema:
            $ref: '#/components/schemas/PreflightFailureEnvelope'
          example: |-
            id: preflight:trace-id:1
            event: preflight.failed
            data: {"schemaVersion":"1.1","runId":null,"messageId":"7ec1cff4-6705-401c-b36a-692b5f9173a7","sequence":1,"traceId":"trace-id","timestamp":"2026-07-26T08:00:00Z","payload":{"status":"FAILED","error":{"code":"AUTHENTICATION_REQUIRED","message":"请登录后使用智能助手","retryable":false}}}
  schemas:
    AssistantRunRequest:
      type: object
      properties:
        conversationId: {type: string, format: uuid}
        messageId: {type: string, format: uuid}
        query: {type: string, minLength: 1, maxLength: 4000}
        context:
          $ref: '#/components/schemas/RunContext'
      required: [conversationId, messageId, query, context]
      additionalProperties: false
    LangGraphRunRequest:
      type: object
      description: Internal request. Header identity is authoritative and must match user.
      properties:
        conversationId: {type: string, format: uuid}
        messageId: {type: string, format: uuid}
        query: {type: string, minLength: 1, maxLength: 4000}
        context:
          $ref: '#/components/schemas/RunContext'
        user:
          $ref: '#/components/schemas/UserContext'
      required: [conversationId, messageId, query, context, user]
      additionalProperties: false
    UserContext:
      type: object
      properties:
        userId: {type: string, maxLength: 128}
        tenantId: {type: string, maxLength: 128}
        roles:
          type: array
          items: {type: string, maxLength: 64}
          maxItems: 20
          uniqueItems: true
      required: [userId, tenantId, roles]
      additionalProperties: false
    RunContext:
      type: object
      properties:
        locale: {type: string, default: zh-CN, maxLength: 32}
        map:
          $ref: '#/components/schemas/MapContext'
        businessObjectIds:
          type: array
          default: []
          items: {type: string, maxLength: 128}
          maxItems: 50
          uniqueItems: true
      required: [locale]
      additionalProperties: false
    MapContext:
      type: object
      properties:
        visibleLayerIds:
          type: array
          items: {type: integer, minimum: 0, maximum: 5}
          maxItems: 6
          uniqueItems: true
        zoom: {type: number, minimum: 0, maximum: 19}
        extent:
          oneOf:
            - $ref: '#/components/schemas/Extent4326'
            - type: 'null'
      required: [visibleLayerIds]
      additionalProperties: false
    Extent4326:
      type: object
      properties:
        xmin: {type: number}
        ymin: {type: number}
        xmax: {type: number}
        ymax: {type: number}
        wkid: {type: integer, const: 4326}
      required: [xmin, ymin, xmax, ymax, wkid]
      additionalProperties: false
    SseEnvelope:
      type: object
      description: Event data for an accepted Run. The SSE event name selects the payload schema from x-sse-event-schemas.
      properties:
        schemaVersion: {type: string, const: '1.1'}
        runId: {type: string, format: uuid}
        messageId: {type: string, format: uuid}
        sequence: {type: integer, minimum: 1}
        traceId: {type: string, minLength: 1, maxLength: 128}
        timestamp: {type: string, format: date-time}
        payload:
          oneOf:
            - $ref: '#/components/schemas/RunStartedPayload'
            - $ref: '#/components/schemas/RouteSelectedPayload'
            - $ref: '#/components/schemas/RetrievalCompletedPayload'
            - $ref: '#/components/schemas/ToolStartedPayload'
            - $ref: '#/components/schemas/ToolCompletedPayload'
            - $ref: '#/components/schemas/MapResultPayload'
            - $ref: '#/components/schemas/Citation'
            - $ref: '#/components/schemas/AnswerDeltaPayload'
            - $ref: '#/components/schemas/RunCompletedPayload'
            - $ref: '#/components/schemas/RunFailedPayload'
            - $ref: '#/components/schemas/RunCancelledPayload'
      required: [schemaVersion, runId, messageId, sequence, traceId, timestamp, payload]
      additionalProperties: false
    PreflightFailureEnvelope:
      type: object
      description: Failure before a Run exists. Its SSE id is preflight:{traceId}:1 and it is exempt from run.started-first.
      properties:
        schemaVersion: {type: string, const: '1.1'}
        runId: {type: 'null'}
        messageId:
          oneOf:
            - {type: string, format: uuid}
            - {type: 'null'}
        sequence: {type: integer, const: 1}
        traceId: {type: string, minLength: 1, maxLength: 128}
        timestamp: {type: string, format: date-time}
        payload:
          $ref: '#/components/schemas/RunFailedPayload'
      required: [schemaVersion, runId, messageId, sequence, traceId, timestamp, payload]
      additionalProperties: false
    RunStartedPayload:
      type: object
      properties:
        status: {type: string, const: RUNNING}
      required: [status]
      additionalProperties: false
    RouteSelectedPayload:
      type: object
      properties:
        intent: {type: string, enum: [MAP_QUERY, RAG_QA, HYBRID, CLARIFY]}
      required: [intent]
      additionalProperties: false
    RetrievalCompletedPayload:
      type: object
      properties:
        documents: {type: integer, minimum: 0}
      required: [documents]
      additionalProperties: false
    ToolStartedPayload:
      type: object
      properties:
        toolCallId: {type: string, format: uuid}
        toolName: {type: string, enum: [queryMapFeatures, queryMapPoints, queryMapLines, searchHousingCandidates]}
      required: [toolCallId, toolName]
      additionalProperties: false
    ToolCompletedPayload:
      type: object
      properties:
        toolCallId: {type: string, format: uuid}
        toolName: {type: string, enum: [queryMapFeatures, queryMapPoints, queryMapLines, searchHousingCandidates]}
        status: {type: string, enum: [SUCCEEDED, FAILED, REJECTED]}
        durationMs: {type: integer, minimum: 0}
      required: [toolCallId, toolName, status, durationMs]
      additionalProperties: false
    AnswerDeltaPayload:
      type: object
      properties:
        content: {type: string, minLength: 1}
      required: [content]
      additionalProperties: false
    RunCompletedPayload:
      type: object
      properties:
        status: {type: string, const: SUCCEEDED}
        answer: {type: string}
        citations:
          type: array
          items: {$ref: '#/components/schemas/Citation'}
        warnings:
          type: array
          items: {type: string}
      required: [status, answer, citations, warnings]
      additionalProperties: false
    RunFailedPayload:
      type: object
      properties:
        status: {type: string, const: FAILED}
        error: {$ref: '#/components/schemas/ContractError'}
      required: [status, error]
      additionalProperties: false
    RunCancelledPayload:
      type: object
      properties:
        status: {type: string, const: CANCELLED}
        reason: {type: string, maxLength: 128}
      required: [status, reason]
      additionalProperties: false
    MapResultPayload:
      type: object
      properties:
        queryId: {type: string}
        toolCallIds:
          type: array
          minItems: 1
          maxItems: 6
          uniqueItems: true
          items: {type: string, format: uuid}
        mode: {type: string, enum: [replace, append]}
        querySummary: {type: string}
        appliedFilters:
          type: array
          items:
            $ref: '#/components/schemas/AppliedFilter'
        resultSets:
          type: array
          maxItems: 6
          description: Across all resultSets, the sum of features must not exceed 200.
          items:
            $ref: '#/components/schemas/MapResultSet'
        overlays:
          type: array
          maxItems: 20
          description: Polygon display overlays. Return an empty array when no overlays are present.
          items:
            $ref: '#/components/schemas/MapOverlay'
        display:
          $ref: '#/components/schemas/MapDisplay'
        warnings:
          type: array
          items: {type: string}
      required: [queryId, toolCallIds, mode, querySummary, appliedFilters, resultSets, overlays, display, warnings]
      additionalProperties: false
    AppliedFilter:
      type: object
      properties:
        field: {type: string}
        operator: {type: string}
        value: {}
        unit: {type: string}
      required: [field, operator]
      additionalProperties: false
    MapResultSet:
      type: object
      description: returned must equal features.length; total must be greater than or equal to returned.
      properties:
        toolCallId: {type: string, format: uuid}
        role: {type: string, enum: [PRIMARY_RESULTS, HOUSING_CANDIDATES, CONTRIBUTING_ROADS]}
        layerId: {type: integer, minimum: 0, maximum: 5}
        layerName: {type: string}
        geometryType: {type: string, enum: [point, polyline]}
        spatialReference:
          $ref: '#/components/schemas/SpatialReference4326'
        total: {type: integer, minimum: 0}
        returned: {type: integer, minimum: 0, maximum: 200}
        exceededTransferLimit: {type: boolean}
        objectIdField: {type: string}
        features:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/MapFeature'
      required: [toolCallId, role, layerId, layerName, geometryType, spatialReference, total, returned, exceededTransferLimit, objectIdField, features]
      additionalProperties: false
    MapFeature:
      type: object
      properties:
        id: {type: string, pattern: '^[0-5]:.+'}
        attributes: {type: object}
        geometry:
          oneOf:
            - $ref: '#/components/schemas/EsriPoint'
            - $ref: '#/components/schemas/EsriPolyline'
      required: [id, attributes, geometry]
      additionalProperties: false
    EsriPoint:
      type: object
      properties:
        x: {type: number}
        y: {type: number}
        spatialReference:
          $ref: '#/components/schemas/SpatialReference4326'
      required: [x, y, spatialReference]
      additionalProperties: false
    EsriPolyline:
      type: object
      properties:
        paths:
          type: array
          minItems: 1
          items:
            type: array
            minItems: 2
            items:
              type: array
              prefixItems: [{type: number}, {type: number}]
              minItems: 2
              maxItems: 2
        spatialReference:
          $ref: '#/components/schemas/SpatialReference4326'
      required: [paths, spatialReference]
      additionalProperties: false
    EsriPolygon:
      type: object
      properties:
        rings:
          type: array
          minItems: 1
          items:
            type: array
            minItems: 4
            description: Closed ring; the first and last coordinate must be equal.
            items:
              type: array
              prefixItems: [{type: number}, {type: number}]
              minItems: 2
              maxItems: 2
        spatialReference:
          $ref: '#/components/schemas/SpatialReference4326'
      required: [rings, spatialReference]
      additionalProperties: false
    MapOverlay:
      type: object
      properties:
        overlayId: {type: string, pattern: '^buf-[A-Za-z0-9._:-]+$', maxLength: 128}
        kind: {const: ROAD_BUFFER}
        geometryType: {const: polygon}
        spatialReference:
          $ref: '#/components/schemas/SpatialReference4326'
        sourceRoadIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: {type: string, pattern: '^[3-5]:.+'}
        attributes:
          type: object
          properties:
            bufferMeters: {type: integer, minimum: 20, maximum: 2000}
            sourceRoadCount: {type: integer, minimum: 1}
            sourceRoadIdsTruncated: {type: boolean}
          required: [bufferMeters, sourceRoadCount, sourceRoadIdsTruncated]
          additionalProperties: false
        geometry:
          $ref: '#/components/schemas/EsriPolygon'
      required: [overlayId, kind, geometryType, spatialReference, sourceRoadIds, attributes, geometry]
      additionalProperties: false
    SpatialReference4326:
      type: object
      properties:
        wkid: {const: 4326}
      required: [wkid]
      additionalProperties: false
    MapDisplay:
      type: object
      properties:
        fitBounds: {type: boolean}
        paddingPx: {type: integer, minimum: 0, maximum: 200}
        maxZoom: {type: number, minimum: 0, maximum: 19}
        layerOrder:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items: {type: string, enum: [ROAD_BUFFER, CONTRIBUTING_ROADS, HOUSING_CANDIDATES, PRIMARY_RESULTS]}
      required: [fitBounds, paddingPx, maxZoom, layerOrder]
      additionalProperties: false
    Citation:
      type: object
      properties:
        citationId: {type: string}
        ordinal: {type: integer, minimum: 1}
        documentId: {type: string}
        documentVersion: {type: string}
        title: {type: string}
        contentType: {type: string}
        sectionPath:
          type: array
          items: {type: string}
        pageStart: {type: integer, minimum: 1}
        pageEnd: {type: integer, minimum: 1}
        chunkId: {type: string}
        excerpt: {type: string}
        excerptAllowed: {type: boolean}
        score: {type: number, description: Provider-specific retrieval or reranking score; used only for ordering.}
        source:
          type: object
          properties:
            resourceRef: {type: string, minLength: 1, maxLength: 256}
            url: {type: string, pattern: '^/api/assistant/citations/'}
          required: [resourceRef]
          additionalProperties: false
        warnings:
          type: array
          items: {type: string}
      required: [citationId, ordinal, documentId, documentVersion, title, contentType, sectionPath, pageStart, pageEnd, chunkId, excerpt, excerptAllowed, score, source, warnings]
      additionalProperties: false
...
    RunStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
    RunStatusData:
      type: object
      properties:
        runId: {type: string, format: uuid}
        status:
          $ref: '#/components/schemas/RunStatus'
        lastSequence: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        completed:
          oneOf:
            - $ref: '#/components/schemas/RunCompletedPayload'
            - type: 'null'
        error:
          oneOf:
            - $ref: '#/components/schemas/ContractError'
            - type: 'null'
      required: [runId, status, lastSequence, createdAt, updatedAt, completed, error]
      additionalProperties: false
    RunStatusResponse:
      allOf:
        - $ref: '#/components/schemas/ContractSuccess'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RunStatusData'
...
    ContractSuccess:
      type: object
      properties:
        success: {type: boolean, const: true}
        data: {}
        traceId: {type: string}
      required: [success, data, traceId]
      additionalProperties: false
    ContractFailure:
      type: object
      properties:
        success: {type: boolean, const: false}
        error:
          $ref: '#/components/schemas/ContractError'
        traceId: {type: string}
      required: [success, error, traceId]
      additionalProperties: false
    ContractError:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
        retryable: {type: boolean}
        details:
          oneOf:
            - {type: object}
            - {type: 'null'}
      required: [code, message, retryable]
      additionalProperties: false